:white_check_mark: Added test cases for test_github
